+++ liquid/__init__.py
@@ -1,12 +1,14 @@
-__version__ = "0.7.1"
+__version__ = "0.7.2"
 
 from .mode import Mode
 from .filter import Filter
 from .token import Token
-from .context import Context
 from .expression import Expression
-
 from .loaders import FileSystemLoader
+
+from .context import Context
+from .context import Undefined
+from .context import StrictUndefined
 
 from .environment import Environment
 from .environment import Template